websrv-flask/app/requirements.txt
python-memcached
pymongo
argon2
Flask-Babel
pyyaml
Flask-SQLAlchemy
sqlalchemy_utils
psycopg2
deprecated
